Core: extract reusable csf-tools from react to csf-tools/story-shape
