+++ code/core/src/csf-tools/index.ts
@@ -3,6 +3,7 @@ export * from './ConfigFile.ts';
 export * from './getStorySortParameter.ts';
 export * from './jsdoc.ts';
 export * from './enrichCsf.ts';
+export * from './story-shape/index.ts';
 export { babelParse } from 'storybook/internal/babel';
 export { vitestTransform } from './vitest-plugin/transformer.ts';
 export { componentTransform } from './vitest-plugin/component-transformer.ts';
